# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bankit fintech API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change user password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Users ──────────────────────────────────────────────────
...
  /accounts:
    get:
      operationId: listBankAccounts
      tags: [accounts]
      summary: List connected bank accounts (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bank accounts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [accounts]
      summary: Connect a new bank account (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "201":
          description: Bank account connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /accounts/{id}:
    patch:
      operationId: updateBankAccount
      tags: [accounts]
      summary: Update bank account (label, enable/disable)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
    delete:
      operationId: deleteBankAccount
      tags: [accounts]
      summary: Disconnect a bank account (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Beneficiaries ──────────────────────────────────────────
  /beneficiaries:
    get:
      operationId: listBeneficiaries
      tags: [beneficiaries]
      summary: List beneficiaries (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Beneficiaries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Beneficiary"
    post:
      operationId: createBeneficiary
      tags: [beneficiaries]
      summary: Add a beneficiary (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeneficiaryInput"
      responses:
        "201":
          description: Beneficiary added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beneficiary"
  /beneficiaries/{id}:
    delete:
      operationId: deleteBeneficiary
      tags: [beneficiaries]
      summary: Remove a beneficiary (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Reports ────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ────────────────────────────────────────────
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8

    # ── Users ───────────────────────────────────────────
    User:
      type: object
      required: [id, email, name, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, account_officer, sales_rep]
        isActive:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
...
    BankAccount:
      type: object
      required: [id, label, accountNumber, bankName, isActive]
      properties:
        id:
          type: integer
        label:
          type: string
        accountNumber:
          type: string
        bankName:
          type: string
        isActive:
          type: boolean
        balance:
          type: ["number", "null"]
        currency:
          type: string
    BankAccountInput:
      type: object
      required: [label, accountNumber, bankName]
      properties:
        label:
          type: string
        accountNumber:
          type: string
        bankName:
          type: string
        currency:
          type: string
    BankAccountUpdate:
      type: object
      properties:
        label:
          type: string
        isActive:
          type: boolean

    # ── Beneficiaries ────────────────────────────────────
    Beneficiary:
      type: object
      required: [id, name, accountNumber, bankName]
      properties:
        id:
          type: integer
        name:
          type: string
        accountNumber:
          type: string
        bankName:
          type: string
        email:
          type: ["string", "null"]
    BeneficiaryInput:
      type: object
      required: [name, accountNumber, bankName]
      properties:
        name:
          type: string
        accountNumber:
          type: string
        bankName:
          type: string
        email:
          type: string

    # ── Reports ──────────────────────────────────────────
...