# lib/api-zod/src/generated/types/bankAccountUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bankit fintech API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BankAccountUpdate {
  label?: string;
  isActive?: boolean;
}